// src/types/EActionTypes.ts
export enum EActionTypes {
  FETCH_MOVIES = 'FETCH_MOVIES',
  FETCH_MOVIES_SUCCESS = 'FETCH_MOVIES_SUCCESS',
  FETCH_MOVIES_ERROR = 'FETCH_MOVIES_ERROR',
  FETCH_MOVIE = 'FETCH_MOVIE',
  FETCH_MOVIE_SUCCESS = 'FETCH_MOVIE_SUCCESS',
  FETCH_MOVIE_ERROR = 'FETCH_MOVIE_ERROR',
  FETCH_TRAILER = 'FETCH_TRAILER',
  FETCH_TRAILER_SUCCESS = 'FETCH_TRAILER_SUCCESS',
  FETCH_TRAILER_ERROR = 'FETCH_TRAILER_ERROR',
  FETCH_MOVIE_ACTORS = 'FETCH_MOVIE_ACTORS',
  FETCH_MOVIE_ACTORS_SUCCESS = 'FETCH_MOVIE_ACTORS_SUCCESS',
  FETCH_MOVIE_ACTORS_ERROR = 'FETCH_MOVIE_ACTORS_ERROR',
  SEARCH_MOVIES = 'SEARCH_MOVIES',
  SEARCH_MOVIES_SUCCESS = 'SEARCH_MOVIES_SUCCESS',
  SEARCH_MOVIES_ERROR = 'SEARCH_MOVIES_ERROR',
  FETCH_ACTORS = 'FETCH_ACTORS',
  FETCH_ACTORS_SUCCESS = 'FETCH_ACTORS_SUCCESS',
  FETCH_ACTORS_ERROR = 'FETCH_ACTORS_ERROR',
  FETCH_ACTOR = 'FETCH_ACTOR',
  FETCH_ACTOR_SUCCESS = 'FETCH_ACTOR_SUCCESS',
  FETCH_ACTOR_ERROR = 'FETCH_ACTOR_ERROR',
  SEARCH_ACTORS = 'SEARCH_ACTORS',
  SEARCH_ACTORS_SUCCESS = 'SEARCH_ACTORS_SUCCESS',
  SEARCH_ACTORS_ERROR = 'SEARCH_ACTORS_ERROR',
}